Fix building on the local M1
There were some very weird errors with the new Homebrew 4 [1] update when
running `brew test-bot only-formulae`: No available formula with the
name "elixir". Since Homebrew 4 changed the fetching from "core"
repositories, it seemed like the error was related. The fix here is to
revert to the previous behaviour [2], although we may need to revisit this
in the future if Homebrew deprecates this flag

[1]: https://brew.sh/2023/02/16/homebrew-4.0.0/
[2]: https://github.com/Homebrew/brew/pull/14375